demo/doc-versioning/pyproject.toml
[tool.poetry]
name = "doc-versioning"
version = "0.1.0"
description = "Document (PDF, DOCx, Images) versioning/revisioning and diffing/visualization library"
authors = ["AXA Group Operations S.A."]
readme = 'README.rst'
repository = 'https://github.com/axa-group/Parsr/tree/develop/demo/doc-versioning'

[tool.poetry.dependencies]
python = "^3.8"
parsr-client = "^3.0.0"
flask = "^1.1.2"
pywebview = "^3.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
